fix handling of grace notes to shorten preceding tied notes correctly
In Note_performer::process_music(), when a grace note was encountered,
the immediately preceding Audio_note (or Audio_notes if the grace note
followed a chord) was/were shortened, but it failed to check whether the
Audio_note is part of a tie.  Ensure that any note being shortened in
this way is the head of a tie, if it is part of a tie.

https://code.google.com/p/lilypond/issues/detail?id=3091